Update benchee and add new benchmark
I'm updating to use the version of Benchee to use what's currently on
master so the machine information is included in the benchmark output.
This should make it easier for folks to contribute rather than me
having to always run the benchmarks on my machine.

I'm also adding the thing about string splitting being slow on large
strings since it looks like that's not going to get fixed in Elixir,
but rather they're going to try and fix Erlang, so it might stick
around for a bit.