Reduce test cases by removing "miss" scenarios
To keep the comparisons more focused.

For reference, when the element is not in the list, MapSet.member?/2 is
the fastest for all n. For small n, the difference is almost
insignificant. But for large n, MapSet.member?/2 can be up to 1559.06x
faster than the next fastest (Enum.member?/2) since the other methods
are exhaustive.